Add controller for MetaKG; refactor controller module
