+++ src/controller/__init__.py
@@ -1,0 +1,4 @@
+from .metakg import MetaKGEntity
+from .smartapi import SmartAPIEntity
+
+__all__ = [MetaKGEntity, SmartAPIEntity]
