+++ src/controller/exceptions.py
@@ -1,0 +1,10 @@
+class ControllerError(Exception):
+    pass
+
+
+class NotFoundError(ControllerError):
+    pass
+
+
+class ConflictError(ControllerError):
+    pass
